requirements.txt
python==3.7.7
pytorch==1.9.1
numpy==1.20.3
scipy==1.7.1
pandas==1.3.4
cython==0.29.24